Generated the basic algorithm
First you calculate how many boards can be created with a 1 or 0 in
each position (by just generating every possible board and counting
the 1's and 0's in each position). Next, find the position with the
fewest boards, and set that value as static. It will generate a unique
puzzle in the fastest way. The get_possible_boards function is currently
as naive as possible, and does not yet check if the board generated
is valid.